+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q001_100/q092/Solution1.java
@@ -3,55 +3,53 @@ package com.noob.algorithm.leetcode.q001_100.q092;
 import com.noob.algorithm.base.dataStructure.linkedList.ListNode;
 
 /**
- * 🟡 92 反转链表II
+ * 🟡 092 反转链表II - https://leetcode.cn/problems/reverse-linked-list-ii/description/
  */
 public class Solution1 {
 
     /**
-     * 反转指定范围后的链表
+     * 分段处理：分段 + 反转 + 重新链接
      */
     public ListNode reverseBetween(ListNode head, int left, int right) {
-        // 遍历链表，找到反转位置并截断
-        ListNode cur = head;
-        ListNode preStartNode = null, endNode = null;
-        for (int i = 0; i <= right; i++) {
-            if (i == left - 1) {
-                preStartNode = cur;
-            }
-            if (i == right) {
-                endNode = cur;
-            }
+        // 构建虚拟头节点，基于dummy遍历寻找要进行反转的区间范围相关节点
+        ListNode dummy = new ListNode(-1, head);
+        ListNode cur = dummy;
+        // 获取开始节点的前一个节点
+        for (int i = 0; i < left - 1; i++) {
             cur = cur.next;
         }
-        ListNode startNode = preStartNode.next;
-        // 将链表拆成3段，然后将中间段进行反转后拼接
-        preStartNode.next = null; // 断开连接
-        ListNode endNxtNode = endNode.next; // 记录结束节点的下一个节点链表
-        endNode.next = null; // 断开连接
 
-        // 对startNode这节链表进行反转
+        ListNode preStartNode = cur; // 记录反转起点的前一个节点
+        ListNode startNode = cur.next; // 记录反转起点
+        for (int i = 0; i < right - left + 1; i++) {
+            cur = cur.next;
+        }
+        ListNode endNode = cur; // 记录反转终点
+        ListNode endNxtNode = cur.next; // 记录反转终点的下一个节点
+
+        // ① 断开链表
+        preStartNode.next = null;
+        endNode.next = null;
+        // ② 反转中间段链表
         ListNode reverseNode = reverse(startNode);
-
-        // 重新拼接
+        // ③ 重新链接
         preStartNode.next = reverseNode;
-        startNode.next = endNxtNode;
+        startNode.next = endNxtNode; // 范围反转后原起点变为反转后的链表终点，将其和下一段链表链接在一起
 
-        // 返回拼接后的链表
-        return head;
+        // 返回结果
+        return dummy.next;
     }
 
-    /**
-     * 反转指定链表
-     */
+    // 反转链表
     private ListNode reverse(ListNode head) {
         ListNode pre = null;
         ListNode cur = head;
         while (cur != null) {
-            ListNode nxt = cur.next;
+            ListNode nxt = cur.next; // 记录下一个节点
             cur.next = pre;
-            // 滚动更新指针
-            cur = nxt;
+            // 更新指针
             pre = cur;
+            cur = nxt;
         }
         // 返回反转后的链表
         return pre;
